float로 바꿨던 것들 일단 전부 int로 복구
win api 좌표에서 float를 못 씀

+++ PlantsVsZombies/GameObject/Bullet/Bullet.cpp
@@ -19,7 +19,7 @@ Bullet::~Bullet()
 
 }
 
-void Bullet::Init(float p_ap, float p_ms)
+void Bullet::Init(int p_ap, int p_ms)
 {
 	m_attackPower = p_ap;
 	m_moveSpeed = p_ms;
@@ -35,7 +35,7 @@ void Bullet::Active(Zombie* p_zombie)
 
 }
 
-float Bullet::GetAttackPower()
+int Bullet::GetAttackPower()
 {
 	return m_attackPower;
 }
