+++ PlantsVsZombies/GameObject/Bullet/Bullet.h
@@ -5,17 +5,17 @@
 class Bullet : public PictureBox
 {
 protected:
-	float m_attackPower;
-	float m_moveSpeed;
+	int m_attackPower;
+	int m_moveSpeed;
 
 	void Move();
 public:
 	Bullet(Point p_pos, Size p_size, const wstring& p_imagePath);
 	virtual ~Bullet();
-	void Init(float p_ap, float p_ms);
+	void Init(int p_ap, int p_ms);
 	void Update();
 
 	virtual void Active(Zombie* p_zombie);
 
-	float GetAttackPower();
+	int GetAttackPower();
 };
