+++ 백업/PlantsVsZombies/GameObject/Plant/Plant.h
@@ -4,16 +4,16 @@
 class Plant : public PictureBox
 {
 protected:
-	int m_hp;
+	float m_hp;
 	bool m_isAlive;
 
 	void CheckAlive();
 public:
 	Plant(Point p_pos, Size p_size, const wstring& p_imagePath);
 	virtual ~Plant();
-	virtual void Init(int p_hp);
+	virtual void Init(float p_hp);
 	virtual void Update();
 
-	void TakeDamage(int p_damage);
+	void TakeDamage(float p_damage);
 	bool IsAlive();
 };
